CDSR-1252 Removed duplicate ACC14 display declaration
