+++ app/uk/gov/hmrc/cdsreimbursementclaim/models/eis/claim/BankDetail.scala
@@ -17,13 +17,23 @@
 package uk.gov.hmrc.cdsreimbursementclaim.models.eis.claim
 
 import play.api.libs.json.{Json, OFormat}
+import uk.gov.hmrc.cdsreimbursementclaim.models.claim.{AccountName, AccountNumber, SortCode}
+import uk.gov.hmrc.cdsreimbursementclaim.models.eis.declaration.response.BankAccountDetails
 
 final case class BankDetail(
-  accountHolderName: String,
-  sortCode: String,
-  accountNumber: String
+  accountHolderName: AccountName,
+  sortCode: SortCode,
+  accountNumber: AccountNumber
 )
 
 object BankDetail {
+
+  def from(bankAccountDetails: BankAccountDetails): BankDetail =
+    BankDetail(
+      bankAccountDetails.accountName,
+      bankAccountDetails.sortCode,
+      bankAccountDetails.accountNumber
+    )
+
   implicit val format: OFormat[BankDetail] = Json.format[BankDetail]
 }
