+++ app/uk/gov/hmrc/cdsreimbursementclaim/models/eis/declaration/DisplayResponseDetail.scala
@@ -17,8 +17,7 @@
 package uk.gov.hmrc.cdsreimbursementclaim.models.eis.declaration
 
 import play.api.libs.json.{Json, OFormat}
-import uk.gov.hmrc.cdsreimbursementclaim.models.claim.{AccountDetails, BankDetails, ConsigneeDetails, DeclarantDetails, NdrcDetails}
-import uk.gov.hmrc.cdsreimbursementclaim.models.eis.declaration.response._
+import uk.gov.hmrc.cdsreimbursementclaim.models.eis.declaration.response.{AccountDetails, BankDetails, ConsigneeDetails, DeclarantDetails, NdrcDetails}
 
 final case class DisplayResponseDetail(
   declarationId: String,
@@ -32,7 +31,7 @@ final case class DisplayResponseDetail(
   consigneeDetails: Option[ConsigneeDetails],
   accountDetails: Option[List[AccountDetails]],
   bankDetails: Option[BankDetails],
-  maskedBankDetails: Option[MaskedBankDetails],
+  maskedBankDetails: Option[BankDetails],
   ndrcDetails: Option[List[NdrcDetails]]
 )
